API IN DEVELOPMENT
From this point forward, I will attempt
to write everything using this single
file, otherwise known as the API.  This
should help clean up code, as well as
making global changes MUCH easier.  For
now, the functionality is rather limited.
It can:
 - Take username/password input
 - Gather login credentials
 - Download one page
 - Download multiple pages
 - Write given data to a CSV
In the future, I hope to expand it to not
only reading, but posting as well.  Also
on the roadmap is the ability to read a
forum and extract thread numbers/names,
to make an alternate UI feasible.

Once this is complete, third-party development
should be much easier, and making anything will
require much less effort.  "import aoapi" will
import all the needed commands, and no common
definitions will be necessary.